Extract `useLocale` and `useTableRowColumnCount` hooks for reusability; refactor Issues Table to use them
